backend/requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.9.11
decorator==4.4.1
Django==2.2.7
django-auth-ldap==2.1.0
django-background-tasks==1.2.5
django-cors-headers==3.2.0
django-extensions==2.2.5
filetype==1.0.5
h5py==2.10.0
ipdb==0.12.2
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
llvmlite==0.31.0
matplotlib==3.1.2
numpy==1.17.4
opencv-python>=4.2.0
openpyxl==3.0.5
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pypng==0.0.20
python-decouple==3.3
python-ldap==3.2.0
python-magic==0.4.13
pytz==2019.3
six==1.13.0
scikit-image==0.17.2
scikit-learn==0.23.2
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
Werkzeug==0.16.0
xmltodict==0.12.0

